requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorlog==6.8.2
frozenlist==1.4.1
hikari==2.0.0.dev125
hikari-lightbulb==2.3.5
idna==3.7
multidict==6.0.5
pillow==10.3.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
uvloop==0.19.0
yarl==1.9.4
